Added completed test
changed some of shoppingcartitem class by adding joinColumn attribute
