+++ backend/src/main/java/sg/edu/nus/team3/shoppingcart/model/ShoppingCartItem.java
@@ -4,6 +4,7 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.Getter;
@@ -19,6 +20,7 @@ public class ShoppingCartItem {
     private int id;
 
     @ManyToOne
+    @JoinColumn(name= "shopping_cart_id")
     @Getter @Setter
     private ShoppingCart shoppingCart;
