+++ backend/src/main/java/sg/edu/nus/team3/shoppingcart/service/ShoppingCartService.java
@@ -2,6 +2,7 @@ package sg.edu.nus.team3.shoppingcart.service;
 
 import java.util.ArrayList;
 
+import sg.edu.nus.team3.shoppingcart.model.Product;
 import sg.edu.nus.team3.shoppingcart.model.ShoppingCartItem;
 import sg.edu.nus.team3.shoppingcart.model.User;
 
@@ -11,7 +12,7 @@ public interface ShoppingCartService {
 	// outline the methods but no need to define method body 
 	
 	
-	public void addProductToCart(User id, Product id); 
+	public void addProductToCart(User user_id, Product product_id); 
 	
 	public void updateCartQuantity(int id, int quantity);
